requirements.txt
yfinance
pandas
numpy
requests
matplotlib
apscheduler
beautifulsoup4
pillow
flask
